+++ src/all/kamyroll/AndroidManifest.xml
