+++ src/all/kamyroll/res/mipmap-hdpi/ic_launcher_adaptive_fore.png
