+++ src/all/kamyroll/res/mipmap-mdpi/ic_launcher.png
