+++ src/all/kamyroll/res/mipmap-mdpi/ic_launcher_adaptive_fore.png
