+++ src/all/kamyroll/res/mipmap-xxhdpi/ic_launcher_adaptive_fore.png
